Moved module's security configuration to the common as generic classes. Resolved #1117
