+++ abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/configuration/PlatformModuleConfiguration.java
@@ -21,7 +21,7 @@ import static com.abixen.platform.service.businessintelligence.configuration.Pla
 
 
 @Configuration
-@ComponentScan(basePackages = {CONFIG, CONTROLLER, CHART, FACADE, INTEGRATION, COMMON_CONVERTER})
+@ComponentScan(basePackages = {CONFIG, CONTROLLER, CHART, FACADE, INTEGRATION, COMMON_CONVERTER, SECURITY})
 public class PlatformModuleConfiguration {
 
 }
