--- abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/security/PlatformSecurityExpressionHandler.java
@@ -1,43 +1,0 @@
-/**
- * Copyright (c) 2010-present Abixen Systems. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.abixen.platform.service.businessintelligence.security;
-
-import org.aopalliance.intercept.MethodInvocation;
-import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
-import org.springframework.security.access.expression.method.MethodSecurityExpressionOperations;
-import org.springframework.security.core.Authentication;
-
-
-public class PlatformSecurityExpressionHandler extends DefaultMethodSecurityExpressionHandler {
-
-   // private PageRepository pageRepository;
-
-    public PlatformSecurityExpressionHandler() {
-        super();
-    }
-
-    @Override
-    protected MethodSecurityExpressionOperations createSecurityExpressionRoot(Authentication authentication, MethodInvocation invocation) {
-        PlatformMethodSecurityExpressionRoot root = new PlatformMethodSecurityExpressionRoot(authentication);
-        root.setThis(invocation.getThis());
-        root.setPermissionEvaluator(this.getPermissionEvaluator());
-        //root.setPageRepository(pageRepository);
-        return root;
-    }
-
-    //public void setPageRepository(PageRepository pageRepository){
-        //this.pageRepository = pageRepository;
-    //}
-}
