+++ abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/configuration/PlatformWebContentServiceJpaConfiguration.java
@@ -16,7 +16,7 @@ package com.abixen.platform.service.webcontent.configuration;
 
 import com.abixen.platform.common.configuration.AbstractJpaConfiguration;
 import com.abixen.platform.common.configuration.properties.AbstractPlatformJdbcConfigurationProperties;
-import com.abixen.platform.service.webcontent.security.PlatformAuditorAware;
+import com.abixen.platform.common.security.PlatformAuditorAware;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
