+++ abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/configuration/PlatformWebContentServicePackages.java
@@ -44,4 +44,6 @@ public class PlatformWebContentServicePackages {
     public static final String CONVERTER = MAIN + ".converter";
 
     public static final String COMMON_CONVERTER = COMMON + ".converter";
+
+    public static final String SECURITY = COMMON + ".security";
 }
