--- abixen-platform-web-content-service/src/main/java/com/abixen/platform/service/webcontent/security/PlatformMethodSecurityExpressionRoot.java
@@ -1,60 +1,0 @@
-/**
- * Copyright (c) 2010-present Abixen Systems. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.abixen.platform.service.webcontent.security;
-
-import org.springframework.security.access.expression.SecurityExpressionRoot;
-import org.springframework.security.access.expression.method.MethodSecurityExpressionOperations;
-import org.springframework.security.core.Authentication;
-
-
-public class PlatformMethodSecurityExpressionRoot extends SecurityExpressionRoot implements MethodSecurityExpressionOperations {
-
-    private Object filterObject;
-    private Object returnObject;
-    private Object target;
-
-    public PlatformMethodSecurityExpressionRoot(Authentication authentication) {
-        super(authentication);
-    }
-
-    protected void setThis(Object target) {
-        this.target = target;
-    }
-
-    @Override
-    public void setFilterObject(Object filterObject) {
-        this.filterObject = filterObject;
-    }
-
-    @Override
-    public Object getFilterObject() {
-        return filterObject;
-    }
-
-    @Override
-    public void setReturnObject(Object returnObject) {
-        this.returnObject = returnObject;
-    }
-
-    @Override
-    public Object getReturnObject() {
-        return returnObject;
-    }
-
-    @Override
-    public Object getThis() {
-        return target;
-    }
-}
